+++ packages/vue/src/index.ts
@@ -1,10 +1,8 @@
 export { BlockComponent } from './Block';
 export { LinkComponent, ActiveLink } from './Link';
-export { ReactiveCuriProps } from './interface';
+export { ReactiveResponse } from './interface';
 export { CuriPluginOptions } from './plugin';
 
 import CuriPlugin from './plugin';
-import installCuri from './install';
-import reactiveCuri from './reactive';
 
-export { CuriPlugin, installCuri, reactiveCuri };
+export { CuriPlugin };