+++ packages/vue/src/interface.ts
@@ -2,14 +2,14 @@ import Vue from 'vue';
 import { CuriRouter, Response } from '@curi/core';
 import { Action } from '@hickory/root';
 
-export interface ReactiveCuriProps {
+export interface ReactiveResponse {
   response: Response;
   action: Action;
 }
 
 declare module 'vue/types/vue' {
   interface Vue {
-    $curi: ReactiveCuriProps;
+    $curi: ReactiveResponse;
     $router: CuriRouter;
   }
 }