+++ packages/vue/types/index.d.ts
@@ -1,8 +1,6 @@
 export { BlockComponent } from './Block';
 export { LinkComponent, ActiveLink } from './Link';
-export { ReactiveCuriProps } from './interface';
+export { ReactiveResponse } from './interface';
 export { CuriPluginOptions } from './plugin';
 import CuriPlugin from './plugin';
-import installCuri from './install';
-import reactiveCuri from './reactive';
-export { CuriPlugin, installCuri, reactiveCuri };
+export { CuriPlugin };